Prevent memory impl to access private fields + add missing pub/accessors (#80)
* prevent memory impl to access private fields + add missing pub/accessors

* move stack executor in its own submodule
