+++ src/executor/mod.rs
@@ -2,13 +2,7 @@
 //!
 //! Executors are structs that hook gasometer and the EVM core together. It
 //! also handles the call stacks in EVM.
+//!
+//! Currently only a stack-based (customizable) executor is provided.
 
-mod stack;
-
-pub use self::stack::{
-	MemoryStackAccount, MemoryStackState, MemoryStackSubstate, PrecompileFailure, PrecompileFn,
-	PrecompileOutput, PrecompileSet, StackExecutor, StackExitKind, StackState,
-	StackSubstateMetadata,
-};
-
-pub use ethereum::Log;
+pub mod stack;
